% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\docType{methods}
\name{CNV.check}
\alias{CNV.check}
\alias{CNV.check,CNV.data-method}
\title{CNV.check}
\usage{
CNV.check(object)

\S4method{CNV.check}{CNV.data}(object)
}
\arguments{
\item{object}{\code{CNV.data} object.}
}
\value{
\code{CNV.data} object.
}
\description{
Check intensity values.
}
\details{
This method checks if intensities are positive and not NA. If not, they are set to 1. Warnings are given if intensities are abnormally high or low (> 50000 or < 5000, respectively).
}
\author{
Volker Hovestadt \email{conumee@hovestadt.bio}
}

