% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation.R
\name{CNV.create_bins}
\alias{CNV.create_bins}
\title{CNV.create_bins}
\usage{
CNV.create_bins(hg19.anno, bin_minsize = 50000, hg19.gap, hg19.exclude)
}
\arguments{
\item{hg19.anno}{foo}

\item{bin_minsize}{foo}

\item{hg19.gap}{foo}

\item{hg19.exclude}{foo}
}
\value{
\code{GRanges} object.
}
\description{
Split genome into bins of defined size.
}

