% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{CNV.import}
\alias{CNV.import}
\title{CNV.import}
\usage{
CNV.import(seg, genome = "hg19", name = "ID", score = "seg.median")
}
\arguments{
\item{seg}{the .seg filename}

\item{genome}{genome against which segments were annotated ("hg19")}

\item{name}{.seg file column to use as $name metadata ("ID")}

\item{score}{.seg file column to use as $score metadata ("seg.median")}
}
\value{
\code{GRanges} class.
}
\description{
import segments where !is.na(pval) from a .seg file into GRanges
}
\details{
This makes the output of CNV.write easier to merge and annotate.
}
\examples{

os <- system.file("extdata", "osteo.hg19.seg", package="conumee")
osteo <- CNV.import(os, genome="hg19")

nbl <- system.file("extdata", "neuro.hg19.seg", package="conumee")
neuro <- CNV.import(nbl, genome="hg19") 

merged <- sort(c(osteo, neuro))
show(merged)

# library(Homo.sapiens)
# library(VariantAnnotation)
# annotated <- locateVariants(merged, Homo.sapiens)
# show(annotated)

}
\author{
Tim Triche, Jr.\email{tim.triche@gmail.com}
}
