% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation.R
\name{CNV.merge_bins}
\alias{CNV.merge_bins}
\title{CNV.merge_bins}
\usage{
CNV.merge_bins(hg19.anno, hg19.tile, bin_minprobes = 20, hg19.probes,
  bin_maxsize = 5e+06, verbose = FALSE)
}
\arguments{
\item{hg19.anno}{foo}

\item{hg19.tile}{foo}

\item{bin_minprobes}{foo}

\item{hg19.probes}{foo}

\item{bin_maxsize}{foo}

\item{verbose}{foo}
}
\value{
\code{GRanges} object.
}
\description{
Merge bins containing less than the defined number probes with neighboring bin containing fewer probes.
}

