% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{CNV.processsummary}
\alias{CNV.processsummary}
\title{CNV.processsummary}
\usage{
CNV.processsummary(object, ...)
}
\arguments{
\item{object}{\code{CNV.data} object.}

\item{pheno=NULL, }{data.frame phenosheet containing info about samples with columns as seen in other parameters}

\item{labels=NULL, }{charcter. labels column in pheno sheet}

\item{interest_groups=NULL, }{character specify the labels/groups you are interested in analysing (needs to be column name of phenosheet)}

\item{identifier=NULL, }{character. (phenosheet-)column mapping CNV data (from names()) to this phenosheet column}

\item{controls=NULL, }{character. labelname of controls in labels column}

\item{anno=NULL, }{CNV.anno object. annotation object from CNV.create_anno}

\item{cutoff=0.4, }{numeric. threshold value to call gain or loss of glFrequency of DNAcopy package}

\item{summary_plots=TRUE, }{logical. whether to perform summary plots}

\item{sample_plots=TRUE, }{logical. whether to perform single sample plots (by now all single sample methods will be performed from the conumee package)}

\item{summaryanalysis}{CNV.summaryanalysis object as returned from CNV.processsummary function i.e.}

\item{intensity_vals.}{intensityvalues for the groups of interest specified as matrices as lists.}

\item{cnv_seg_data}{CNV.segment sample data as list grouped by groups of interest.}

\item{gl_freqs}{gl_frequency objects of the DNAcopy package for selected groups as lists.}

\item{chr}{= "all", character vector. Which chromomsomes to plot. Defaults to 'all'.}

\item{chrX}{= TRUE, logical. Plot values for chrX? Defaults to TRUE. Set CNV.create_anno(chrXY =FALSE)  if chrX and Y should not be included at all.}

\item{chrY}{= TRUE,  logical. Plot values for chrY? Defaults to TRUE}

\item{centromere}{= TRUE, logical. Show dashed lines at centromeres? Defaults to TRUE.}

\item{main}{= NULL, character. Title of the plot. Defaults to interest groups names.}

\item{ylim}{= c(-1, 1) numeric vector. The y limits of the plot. Defaults to c(-1, 1).}

\item{set_par}{logical. Use recommended graphical parameters for \code{oma} and \code{mar}? Defaults to \code{TRUE}. Original parameters are restored afterwards.}

\item{save=}{TRUE, logical. whether to save to path or not}

\item{path=}{NULL character. path to savings without trailing / !}
}
\value{
\code{CNV.summaryanalysis} object.
}
\description{
Create a CNV.summaryanalysis object containing grouped cnv analysis data with option of summaryplotting or single sample analysis and plots.
}
\details{
This method generates a CNV.summaryanalyis object needed to generate summary plots with CNV.summaryplot of specified interest groups. On the y-axis gains and losses are plotted according to group frequency. See parameters for more information.
}
\examples{
# see CNV.summaryanalisys for an example
}
\author{
Samir Jabari\email{samir.jabari@fau.de}
}
