% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{CNV.summaryanalysis-class}
\alias{CNV.summaryanalysis-class}
\alias{show,CNV.summaryanalysis-method}
\alias{names,CNV.summaryanalysis-method}
\alias{names<-,CNV.summaryanalysis-method}
\title{CNV.summaryanalisys class}
\usage{
\S4method{show}{CNV.summaryanalysis}(object)

\S4method{names}{CNV.summaryanalysis}(x)

\S4method{names}{CNV.summaryanalysis}(x) <- value
}
\arguments{
\item{object}{\code{CNV.summaryanalysis} object}

\item{x}{\code{CNV.summaryanalysis} object (defined by \code{show} generic).}

\item{value}{Replacement names.}
}
\value{
\code{CNV.summaryanalysis} class.
}
\description{
CNV summary analysis of grouped data of samples is stored in this class
}
\details{
Use \code{CNV.processsummary} to create.
}
\examples{
# load RGset #
RGset<-readRDS('path/to/RGset.rds')
### or use an Mset instead ###

# process to Mset #
Mset<-preprocessIllumina(RGset)

# load conumee package and load via CNV.load method #
library(conumee)
Data.data<- CNV.load(Mset)
# create annotation object; not complete and non working example below!! --> for reference and help read conumee vignette #
anno <- CNV.create_anno(array_type = array_type, detail_regions = detail_regions)


# fetch controls from your Data #
Data.controls<- grep( 'Control', names(Data.data))

# read in phenosheet of samples; an identifier column mapping CNV data to phenosheet (named ID here ) should be specified #
test_pheno=read.csv(file ='../test_pheno.csv')
keep <- match(test_pheno$ID, names(Data.data))
Data.data_test<-Data.data[c(keep)]

x<-CNV.processsummary(object=Data.data_test,
                                pheno=test_pheno,
                                labels='labels',
                                interest_groups=NULL,
                                identifier='ID',
                                controls=c('Control'),
                                anno= anno,
                                summary_plots=TRUE,
                                sample_plots=TRUE,
                                chr = "all",
                                chrX = TRUE,
                                chrY = TRUE,
                                centromere = TRUE,
                                main = NULL,
                                ylim = c(-1, 1),
                                set_par = TRUE,
                                save=TRUE,
                                path="Path/to/savings")

# general information
x
show(x)

# show or replace sample name
names(x)
names(x) <- 'Sample 1'

}
\author{
Samir Jabari \email{samir.jabari@fau.de}
}
