% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\name{CNV.summaryplot}
\alias{CNV.summaryplot}
\alias{CNV.summaryplot,CNV.summaryanalysis-method}
\title{CNV.summaryplot}
\usage{
CNV.summaryplot(object, ...)

\S4method{CNV.summaryplot}{CNV.summaryanalysis}(
  object = NULL,
  chr = "all",
  chrX = TRUE,
  chrY = TRUE,
  centromere = TRUE,
  main = NULL,
  ylim = c(-1, 1),
  set_par = TRUE,
  save = TRUE,
  path = NULL
)
}
\arguments{
\item{object}{\code{CNV.summaryanalysis} object.}

\item{chr}{= "all", character vector. Which chromomsomes to plot. Defaults to 'all'.}

\item{chrX}{= TRUE, logical. Plot values for chrX? Defaults to TRUE. Set CNV.create_anno(chrXY =FALSE)  if chrX and Y should not be included at all.}

\item{chrY}{= TRUE,  logical. Plot values for chrY? Defaults to TRUE}

\item{centromere}{= TRUE, logical. Show dashed lines at centromeres? Defaults to TRUE.}

\item{main}{= NULL, character. Title of the plot. Defaults to interest groups names.}

\item{ylim}{= c(-1, 1) numeric vector. The y limits of the plot. Defaults to c(-1, 1).}

\item{set_par}{logical. Use recommended graphical parameters for \code{oma} and \code{mar}? Defaults to \code{TRUE}. Original parameters are restored afterwards.}

\item{summaryanalysis}{CNV.summaryanalysis object as returned from CNV.processsummary function i.e.}

\item{intensity_vals.}{intensityvalues for the groups of interest specified as matrices as lists.}

\item{cnv_seg_data}{CNV.segment sample data as list grouped by groups of interest.}

\item{gl_freqs}{gl_frequency objects of the DNAcopy package for selected groups as lists.}

\item{save=}{TRUE, logical. whether to save to path or not}

\item{path=}{NULL character. path to savings without trailing / !}
}
\value{
\code{NULL}.
}
\description{
Create CNV summary plots for Samples already processed with CNV.segment or via CNV.processsummary.
}
\details{
This method provides the functionality for generating summary CNV plots of specified interest groups. On the y-axis gains and losses are plotted according to group frequency. See parameters for more information.
}
\examples{
# see CNV.summaryanalisys for an example
}
\author{
Samir Jabari\email{samir.jabari@fau.de}
}
