% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{detail_regions}
\alias{detail_regions}
\title{detail_regions}
\description{
Example of genomic regions to be analyzed in detail (e.g. candidate oncogenes/TSGs).
}
\details{
Imported using \code{rtracklayer}. Raw data stored in \code{inst/extdata/detail_regions.bed}.
}
\author{
Volker Hovestadt \email{conumee@hovestadt.bio}
}
