% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{exclude_regions}
\alias{exclude_regions}
\title{exclude_regions}
\description{
Example of genomic regions to exclude (e.g. known polymorphic regions).
}
\details{
Imported using \code{rtracklayer}. Raw data stored in \code{inst/extdata/exclude_regions.bed}.
}
\author{
Volker Hovestadt \email{conumee@hovestadt.bio}
}

