% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{read.450k.url}
\alias{read.450k.url}
\title{read.450k.url}
\usage{
read.450k.url(url = NULL, idat = NULL)
}
\arguments{
\item{url}{URL of the directory in which the IDAT files are located.}

\item{idat}{Vector of IDAT names. \code{url} and \code{idat} default to the TCGA example described in the vignette.}
}
\value{
\code{RGChannelSet} object.
}
\description{
Read IDAT files from the web.
}
\details{
This method downloads the provided list of IDAT files to a temporary folder (using the \code{RCurl} package). It then uses the `read.450k.exp` method of the `minfi` package.
}
\examples{
RGsetTCGA <- read.450k.url()
}
\author{
Volker Hovestadt \email{conumee@hovestadt.bio}
}

