% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{tbl_ucsc}
\alias{tbl_ucsc}
\title{tbl_ucsc}
\description{
UCSC tables required for creating annotation object.
}
\details{
Imported using \code{rtracklayer::browserSession('UCSC')}: \code{chromInfo}, \code{gap}, \code{cytoBand}.
}
\author{
Volker Hovestadt \email{conumee@hovestadt.bio}
}

