% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{tcgaBRCA.sentrix2name}
\alias{tcgaBRCA.sentrix2name}
\title{tcgaBRCA.sentrix2name}
\description{
Named vector for Sentrix ID to TCGA ID conversion of breast cancer example data (see README).
}
\details{
Based on \code{https://tcga-data.nci.nih.gov/tcgafiles/ftp_auth/distro_ftpusers/anonymous/tumor/brca/cgcc/jhu-usc.edu/humanmethylation450/methylation/jhu-usc.edu_BRCA.HumanMethylation450.aux.1.8.0/BRCA.mappings.csv}.
}
\author{
Volker Hovestadt \email{conumee@hovestadt.bio}
}
